@@ -1,5 +1,6 @@
 // Copyright 2020 Damir Jelić
 // Copyright 2020 The Matrix.org Foundation C.I.C.
+// Copyright 2022 Famedly GmbH
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -19,7 +20,10 @@ use std::{
     future::Future,
     io::Read,
     pin::Pin,
-    sync::{Arc, RwLock as StdRwLock},
+    sync::{
+        atomic::{AtomicU64, Ordering::SeqCst},
+        Arc, RwLock as StdRwLock,
+    },
 };
 
 use anymap2::any::CloneAnySendSync;
@@ -80,7 +84,10 @@ use crate::{
     attachment::{AttachmentInfo, Thumbnail},
     config::RequestConfig,
     error::{HttpError, HttpResult},
-    event_handler::{EventHandler, EventHandlerData, EventHandlerResult, EventKind, SyncEvent},
+    event_handler::{
+        EventHandler, EventHandlerData, EventHandlerHandle, EventHandlerResult,
+        EventHandlerWrapper, EventKind, SyncEvent,
+    },
     http_client::HttpClient,
     room, Account, Error, Result,
 };
@@ -101,8 +108,8 @@ const DEFAULT_UPLOAD_SPEED: u64 = 125_000;
 const MIN_UPLOAD_REQUEST_TIMEOUT: Duration = Duration::from_secs(60 * 5);
 
 type EventHandlerFut = Pin<Box<dyn Future<Output = ()> + Send>>;
-type EventHandlerFn = Box<dyn Fn(EventHandlerData<'_>) -> EventHandlerFut + Send + Sync>;
-type EventHandlerMap = BTreeMap<(EventKind, &'static str), Vec<EventHandlerFn>>;
+pub(crate) type EventHandlerFn = dyn Fn(EventHandlerData<'_>) -> EventHandlerFut + Send + Sync;
+type EventHandlerMap = BTreeMap<(EventKind, &'static str), Vec<EventHandlerWrapper>>;
 
 type NotificationHandlerFut = EventHandlerFut;
 type NotificationHandlerFn =
@@ -156,6 +163,9 @@ pub(crate) struct ClientInner {
     event_handlers: RwLock<EventHandlerMap>,
     /// Custom event handler context. See `register_event_handler_context`.
     event_handler_data: StdRwLock<AnyMap>,
+    /// When registering a event handler, the current value is used for the
+    /// handlers identification, then the counter is incremented.
+    event_handler_counter: AtomicU64,
     /// Notification handlers. See `register_notification_handler`.
     notification_handlers: RwLock<Vec<NotificationHandlerFn>>,
     /// Whether the client should operate in application service style mode.
@@ -346,9 +356,11 @@ impl Client {
     /// "context" arguments: They have to implement [`EventHandlerContext`].
     /// This trait is named that way because most of the types implementing it
     /// give additional context about an event: The room it was in, its raw form
-    /// and other similar things. As an exception to this,
-    /// [`Client`] also implements the `EventHandlerContext` trait
-    /// so you don't have to clone your client into the event handler manually.
+    /// and other similar things. As two exceptions to this,
+    /// [`Client`] and [`EventHandlerHandle`] also implement the
+    /// `EventHandlerContext` trait so you don't have to clone your client
+    /// into the event handler manually and a handler can decide to remove
+    /// itself.
     ///
     /// Some context arguments are not universally applicable. A context
     /// argument that isn't available for the given event type will result in
@@ -388,13 +400,15 @@ impl Client {
     /// #     .build()
     /// #     .await
     /// #     .unwrap();
+    ///
     /// client
     ///     .register_event_handler(
     ///         |ev: SyncRoomMessageEvent, room: Room, client: Client| async move {
     ///             // Common usage: Room event plus room and client.
     ///         },
     ///     )
-    ///     .await
+    ///     .await;
+    /// client
     ///     .register_event_handler(
     ///         |ev: SyncRoomMessageEvent, room: Room, encryption_info: Option<EncryptionInfo>| {
     ///             async move {
@@ -403,7 +417,8 @@ impl Client {
     ///             }
     ///         },
     ///     )
-    ///     .await
+    ///     .await;
+    /// client
     ///     .register_event_handler(|ev: SyncRoomTopicEvent| async move {
     ///         // You can omit any or all arguments after the first.
     ///     })
@@ -431,52 +446,116 @@ impl Client {
     ///     move |ev: SyncRoomMessageEvent | {
     ///         let data = data.clone();
     ///         async move {
-    ///             println!("Calling the handler with identifier {}", data);
+    ///             println!("Calling the handler with identifier {data}");
     ///         }
     ///     }
     /// }).await;
     /// # });
     /// ```
-    pub async fn register_event_handler<Ev, Ctx, H>(&self, handler: H) -> &Self
+    pub async fn register_event_handler<Ev, Ctx, H>(&self, handler: H) -> EventHandlerHandle
     where
         Ev: SyncEvent + DeserializeOwned + Send + 'static,
         H: EventHandler<Ev, Ctx>,
         <H::Future as Future>::Output: EventHandlerResult,
     {
-        let event_type = H::ID.1;
-        self.inner.event_handlers.write().await.entry(H::ID).or_default().push(Box::new(
-            move |data| {
-                let maybe_fut = serde_json::from_str(data.raw.get())
-                    .map(|ev| handler.clone().handle_event(ev, data));
+        let event_type = Ev::TYPE;
+        let key = (Ev::KIND, Ev::TYPE);
 
-                Box::pin(async move {
-                    match maybe_fut {
-                        Ok(Some(fut)) => {
-                            fut.await.print_error(event_type);
-                        }
-                        Ok(None) => {
-                            error!(
-                                "Event handler for {} has an invalid context argument",
-                                event_type
-                            );
-                        }
-                        Err(e) => {
-                            warn!(
-                                "Failed to deserialize `{}` event, skipping event handler.\n\
-                                 Deserialization error: {}",
-                                event_type, e,
-                            );
-                        }
+        let handler_fn: Box<EventHandlerFn> = Box::new(move |data| {
+            let maybe_fut = serde_json::from_str(data.raw.get())
+                .map(|ev| handler.clone().handle_event(ev, data));
+
+            Box::pin(async move {
+                match maybe_fut {
+                    Ok(Some(fut)) => {
+                        fut.await.print_error(event_type);
                     }
-                })
-            },
-        ));
+                    Ok(None) => {
+                        error!("Event handler for {} has an invalid context argument", event_type);
+                    }
+                    Err(e) => {
+                        warn!(
+                            "Failed to deserialize `{}` event, skipping event handler.\n\
+                                Deserialization error: {}",
+                            event_type, e,
+                        );
+                    }
+                }
+            })
+        });
 
-        self
+        let handler_id = self.inner.event_handler_counter.fetch_add(1, SeqCst);
+
+        let handle = EventHandlerHandle { handler_id, ev_id: key };
+
+        self.inner
+            .event_handlers
+            .write()
+            .await
+            .entry(key)
+            .or_default()
+            .push(EventHandlerWrapper { handler_fn, handle });
+
+        handle
     }
 
     pub(crate) async fn event_handlers(&self) -> RwLockReadGuard<'_, EventHandlerMap> {
         self.inner.event_handlers.read().await
+    }
+
+    /// Remove the event handler associated with the handle.
+    ///
+    /// Note that handlers that remove themselves will still execute
+    /// with events received in the same sync cycle.
+    ///
+    ///  # Arguments
+    ///
+    /// `handle` - The [`EventHandlerHandle`] that is returned when
+    /// registering the event handler with [`Client::register_event_handler`].
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use futures::executor::block_on;
+    /// # use url::Url;
+    /// # use tokio::sync::mpsc;
+    /// #
+    /// # let homeserver = Url::parse("http://localhost:8080").unwrap();
+    /// #
+    /// use matrix_sdk::{
+    ///     ruma::events::room::member::SyncRoomMemberEvent,
+    ///     Client, event_handler::EventHandlerHandle
+    /// };
+    /// #
+    /// # block_on(async {
+    /// # let client = matrix_sdk::Client::builder()
+    /// #     .homeserver_url(homeserver)
+    /// #     .server_versions([ruma::api::MatrixVersion::V1_0])
+    /// #     .build()
+    /// #     .await
+    /// #     .unwrap();
+    ///
+    /// client
+    ///     .register_event_handler(
+    ///         |ev: SyncRoomMemberEvent, client: Client, handle: EventHandlerHandle| async move {
+    ///             // Common usage: Check arriving Event is the expected one
+    ///             println!("Expected RoomMemberEvent received!");
+    ///             client.remove_event_handler(handle);
+    ///         },
+    ///     )
+    ///     .await;
+    /// # });
+    /// ```
+    pub async fn remove_event_handler(&self, handle: EventHandlerHandle) {
+        let mut event_handlers = self.inner.event_handlers.write().await;
+
+        if let Some(v) = event_handlers.get_mut(&handle.ev_id) {
+            v.retain(|e| e.handle.handler_id != handle.handler_id);
+
+            if v.is_empty() {
+                event_handlers.remove(&handle.ev_id);
+            }
+        }
     }
 
     /// Add an arbitrary value for use as event handler context.
@@ -511,8 +590,8 @@ impl Client {
     /// // Handle used to send messages to the UI part of the app
     /// let my_gui_handle: SomeType = obtain_gui_handle();
     ///
+    /// client.register_event_handler_context(my_gui_handle.clone());
     /// client
-    ///     .register_event_handler_context(my_gui_handle.clone())
     ///     .register_event_handler(
     ///         |ev: SyncRoomMessageEvent, room: Room, gui_handle: Ctx<SomeType>| async move {
     ///             // gui_handle.send(DisplayMessage { message: ev });
@@ -521,12 +600,11 @@ impl Client {
     ///     .await;
     /// # });
     /// ```
-    pub fn register_event_handler_context<T>(&self, ctx: T) -> &Self
+    pub fn register_event_handler_context<T>(&self, ctx: T)
     where
         T: Clone + Send + Sync + 'static,
     {
         self.inner.event_handler_data.write().unwrap().insert(ctx);
-        self
     }
 
     pub(crate) fn event_handler_context<T>(&self) -> Option<T>
@@ -725,7 +803,6 @@ impl Client {
     /// # Example
     ///
     /// ```no_run
-    /// # use std::convert::TryFrom;
     /// # use futures::executor::block_on;
     /// # use url::Url;
     /// # let homeserver = Url::parse("http://example.com").unwrap();
@@ -742,8 +819,8 @@ impl Client {
     ///     .await?;
     ///
     /// println!(
-    ///     "Logged in as {}, got device_id {} and access_token {}",
-    ///     user, response.device_id, response.access_token,
+    ///     "Logged in as {user}, got device_id {} and access_token {}",
+    ///     response.device_id, response.access_token,
     /// );
     /// # anyhow::Ok(()) });
     /// ```
@@ -792,7 +869,6 @@ impl Client {
     /// # Example
     ///
     /// ```no_run
-    /// # use std::convert::TryFrom;
     /// # use matrix_sdk::Client;
     /// # use matrix_sdk::ruma::{assign, DeviceId};
     /// # use futures::executor::block_on;
@@ -1061,7 +1137,6 @@ impl Client {
     /// # Examples
     ///
     /// ```no_run
-    /// # use std::convert::TryFrom;
     /// # use matrix_sdk::Client;
     /// # use matrix_sdk::ruma::{
     /// #     api::client::{
@@ -1092,7 +1167,7 @@ impl Client {
         registration: impl Into<register::v3::Request<'_>>,
     ) -> HttpResult<register::v3::Response> {
         let homeserver = self.homeserver().await;
-        info!("Registering to {}", homeserver);
+        info!("Registering to {homeserver}");
 
         let config = if self.inner.appservice_mode {
             Some(RequestConfig::short_retry().force_auth())
@@ -1224,7 +1299,6 @@ impl Client {
     /// # Examples
     /// ```no_run
     /// use matrix_sdk::Client;
-    /// # use std::convert::TryInto;
     /// # use url::Url;
     /// # let homeserver = Url::parse("http://example.com").unwrap();
     /// # let limit = Some(10);
@@ -1300,7 +1374,6 @@ impl Client {
     /// # Examples
     ///
     /// ```no_run
-    /// # use std::convert::TryFrom;
     /// # use url::Url;
     /// # use matrix_sdk::Client;
     /// # use futures::executor::block_on;
@@ -1415,7 +1488,6 @@ impl Client {
     /// # use matrix_sdk::{Client, config::SyncSettings};
     /// # use futures::executor::block_on;
     /// # use url::Url;
-    /// # use std::convert::TryFrom;
     /// # block_on(async {
     /// # let homeserver = Url::parse("http://localhost:8080")?;
     /// # let mut client = Client::new(homeserver).await?;
@@ -1498,7 +1570,6 @@ impl Client {
     /// # use matrix_sdk::{Client, config::SyncSettings};
     /// # use futures::executor::block_on;
     /// # use url::Url;
-    /// # use std::convert::TryFrom;
     /// # block_on(async {
     /// # let homeserver = Url::parse("http://localhost:8080")?;
     /// # let mut client = Client::new(homeserver).await?;
@@ -1546,7 +1617,7 @@ impl Client {
     /// # use futures::executor::block_on;
     /// # use serde_json::json;
     /// # use url::Url;
-    /// # use std::{collections::BTreeMap, convert::TryFrom};
+    /// # use std::collections::BTreeMap;
     /// # block_on(async {
     /// # let homeserver = Url::parse("http://localhost:8080")?;
     /// # let mut client = Client::new(homeserver).await?;
@@ -1838,8 +1909,6 @@ impl Client {
                 if callback(r).await == LoopCtrl::Break {
                     return;
                 }
-            } else {
-                continue;
             }
 
             Client::delay_sync(&mut last_sync_time).await
