@@ -36,7 +36,7 @@ impl Client {
         for (room_id, room_info) in &rooms.join {
             let room = self.get_room(room_id);
             if room.is_none() {
-                error!("Can't call event handler, room {} not found", room_id);
+                error!("Can't call event handler, room {room_id} not found");
                 continue;
             }
 
@@ -53,7 +53,7 @@ impl Client {
         for (room_id, room_info) in &rooms.leave {
             let room = self.get_room(room_id);
             if room.is_none() {
-                error!("Can't call event handler, room {} not found", room_id);
+                error!("Can't call event handler, room {room_id} not found");
                 continue;
             }
 
@@ -68,7 +68,7 @@ impl Client {
         for (room_id, room_info) in &rooms.invite {
             let room = self.get_room(room_id);
             if room.is_none() {
-                error!("Can't call event handler, room {} not found", room_id);
+                error!("Can't call event handler, room {room_id} not found");
                 continue;
             }
 
@@ -88,7 +88,7 @@ impl Client {
                 let room = match self.get_room(room_id) {
                     Some(room) => room,
                     None => {
-                        warn!("Can't call notification handler, room {} not found", room_id);
+                        warn!("Can't call notification handler, room {room_id} not found");
                         continue;
                     }
                 };
@@ -128,9 +128,7 @@ impl Client {
                 Ok(r)
             }
             Err(e) => {
-                error!("Received an invalid response: {}", e);
-                Self::sleep().await;
-
+                error!("Received an invalid response: {e}");
                 Err(e)
             }
         }
