@@ -14,7 +14,6 @@
 
 use std::{
     collections::{BTreeMap, HashMap},
-    convert::TryInto,
     fmt,
     ops::Deref,
     sync::{
@@ -577,9 +576,8 @@ impl ReadOnlyAccount {
             // so.
             if count != old_count {
                 debug!(
-                    "Updated uploaded one-time key count {} -> {}.",
+                    "Updated uploaded one-time key count {} -> {count}.",
                     self.uploaded_key_count(),
-                    count
                 );
             }
 
