@@ -13,7 +13,6 @@
 // limitations under the License.
 
 use std::{
-    convert::{TryFrom, TryInto},
     matches,
     sync::{Arc, Mutex},
     time::Duration,
@@ -546,10 +545,8 @@ impl SasState<Started> {
             let commitment = calculate_commitment(our_public_key, content);
 
             info!(
-                "Calculated commitment for pubkey {} and content {:?} {}",
+                "Calculated commitment for pubkey {} and content {content:?} {commitment}",
                 our_public_key.to_base64(),
-                content,
-                commitment
             );
 
             if let Ok(accepted_protocols) = AcceptedProtocols::try_from(method_content) {
@@ -1224,8 +1221,6 @@ impl SasState<Cancelled> {
 
 #[cfg(test)]
 mod tests {
-    use std::convert::TryFrom;
-
     use matrix_sdk_test::async_test;
     use ruma::{
         device_id,
