@@ -165,7 +165,6 @@ impl Common {
     ///
     /// # Examples
     /// ```no_run
-    /// # use std::convert::TryFrom;
     /// use matrix_sdk::{room::MessagesOptions, Client};
     /// # use matrix_sdk::ruma::{
     /// #     api::client::filter::RoomEventFilter,
@@ -263,7 +262,6 @@ impl Common {
     ///
     /// # Examples
     /// ```no_run
-    /// # use std::convert::TryFrom;
     /// use matrix_sdk::Client;
     /// # use matrix_sdk::ruma::{
     /// #     api::client::filter::RoomEventFilter,
@@ -315,8 +313,10 @@ impl Common {
             for await item in backward_store {
                 match item {
                     Ok(event) => yield Ok(event),
-                    Err(TimelineStreamError::EndCache { fetch_more_token }) => if let Err(error) = room.request_messages(&fetch_more_token).await {
-                        yield Err(error);
+                    Err(TimelineStreamError::EndCache { fetch_more_token }) => {
+                        if let Err(error) = room.request_messages(&fetch_more_token).await {
+                            yield Err(error);
+                        }
                     },
                     Err(TimelineStreamError::Store(error)) => yield Err(error.into()),
                 }
@@ -343,7 +343,6 @@ impl Common {
     ///
     /// # Examples
     /// ```no_run
-    /// # use std::convert::TryFrom;
     /// use matrix_sdk::Client;
     /// # use matrix_sdk::ruma::{
     /// #     api::client::filter::RoomEventFilter,
@@ -402,7 +401,6 @@ impl Common {
     ///
     /// # Examples
     /// ```no_run
-    /// # use std::convert::TryFrom;
     /// use matrix_sdk::Client;
     /// # use matrix_sdk::ruma::{
     /// #     api::client::filter::RoomEventFilter,
@@ -445,8 +443,10 @@ impl Common {
             for await item in backward_store {
                 match item {
                     Ok(event) => yield Ok(event),
-                    Err(TimelineStreamError::EndCache { fetch_more_token }) => if let Err(error) = room.request_messages(&fetch_more_token).await {
-                        yield Err(error);
+                    Err(TimelineStreamError::EndCache { fetch_more_token }) => {
+                        if let Err(error) = room.request_messages(&fetch_more_token).await {
+                            yield Err(error);
+                        }
                     },
                     Err(TimelineStreamError::Store(error)) => yield Err(error.into()),
                 }
@@ -1075,7 +1075,9 @@ impl Common {
 
 /// Options for [`messages`][Common::messages].
 ///
-/// See that method and <https://spec.matrix.org/v1.3/client-server-api/#get_matrixclientv3roomsroomidmessages> for details.
+/// See that method and
+/// <https://spec.matrix.org/v1.3/client-server-api/#get_matrixclientv3roomsroomidmessages>
+/// for details.
 #[derive(Debug)]
 #[non_exhaustive]
 pub struct MessagesOptions<'a> {
